Make ProgramService a class
